+++ components/bank/BankInfoFooter.vue
